investing/requirements.txt
# This file is autogenerated by pip-compile with Python 3.11
anyio==3.7.1
black==23.11.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
idna==3.6
iniconfig==2.0.0
mypy==1.7.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.1.0
pluggy==1.3.0
pydantic==2.5.0
pydantic-core==2.14.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
ruff==0.1.6
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.9.0
urllib3==2.1.0
uvicorn==0.24.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
